Split WalletView into component and container files
