+++ ui/app/components/app/wallet-view/index.js
@@ -1,0 +1,1 @@
+export { default } from './wallet-view.container'
